requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
tensorflow-cpu